+++ example-dagster-pipes-rust-project/rust_processing_jobs/src/main.rs
@@ -1,8 +1,8 @@
-use dagster_pipes_rust::{open_dagster_pipes, AssetCheckSeverity};
+use dagster_pipes_rust::{open_dagster_pipes, AssetCheckSeverity, DagsterPipesError};
 use serde_json::json;
 
-fn main() {
-    let mut context = open_dagster_pipes();
+fn main() -> Result<(), DagsterPipesError> {
+    let mut context = open_dagster_pipes()?;
     // See supported metadata types here:
     // https://github.com/dagster-io/dagster/blob/master/python_modules/dagster/dagster/_core/pipes/context.py#L133
     let metadata = json!({"row_count": {"raw_value": 100, "type": "int"}});
@@ -14,4 +14,5 @@ fn main() {
         AssetCheckSeverity::Warn,
         json!({"quality": {"raw_value": 5, "type": "int"}}),
     );
+    Ok(())
 }