+++ src/lib.rs
@@ -6,9 +6,12 @@ use std::collections::HashMap;
 use std::fs::OpenOptions;
 use std::io::Write;
 
+use context_loader::PayloadErrorKind;
+use params_loader::ParamsError;
 use serde::{Deserialize, Serialize};
 use serde_json::json;
 use serde_json::Value;
+use thiserror::Error;
 
 use crate::context_loader::PipesContextLoader;
 use crate::context_loader::PipesDefaultContextLoader;
@@ -93,15 +96,28 @@ struct PipesMessagesParams {
     stdio: Option<String>, // stderr | stdout (unsupported)
 }
 
+#[derive(Debug, Error)]
+#[non_exhaustive]
+pub enum DagsterPipesError {
+    #[error("dagster pipes failed to load params: {0}")]
+    #[non_exhaustive]
+    ParamsLoader(#[from] ParamsError),
+
+    #[error("dagster pipes failed to load context: {0}")]
+    #[non_exhaustive]
+    ContextLoader(#[from] PayloadErrorKind),
+}
+
 // partial translation of
 // https://github.com/dagster-io/dagster/blob/258d9ca0db/python_modules/dagster-pipes/dagster_pipes/__init__.py#L798-L838
-pub fn open_dagster_pipes() -> PipesContext {
+pub fn open_dagster_pipes() -> Result<PipesContext, DagsterPipesError> {
     let params_loader = PipesEnvVarParamsLoader::new();
     let context_loader = PipesDefaultContextLoader::new();
 
-    let context_params = params_loader.load_context_params();
-    let message_params = params_loader.load_message_params();
+    let context_params = params_loader.load_context_params()?;
+    let context_data = context_loader.load_context(context_params)?;
 
+    let message_params = params_loader.load_message_params()?;
     // TODO: Refactor into MessageWriter impl
     let path = match &message_params["path"] {
         Value::String(string) => string.clone(),
@@ -112,8 +128,8 @@ pub fn open_dagster_pipes() -> PipesContext {
     //    panic!("only stderr supported for dagster pipes messages")
     //}
 
-    PipesContext {
-        data: context_loader.load_context(context_params),
+    Ok(PipesContext {
+        data: context_data,
         writer: PipesFileMessageWriter { path },
-    }
+    })
 }